error handling in repo

+++ models/users.go
@@ -1,7 +1,7 @@
 package models
 
 type Users struct {
-    ID        int    `json:"user_id"`
+    ID        int64    `json:"user_id"`
     Username  string `json:"username"`
     Password  string `json:"password"`
     Role      string `json:"role"` // e.g., "patient", "admin"
